synced_folder: convert to using slave ssh mount by default
This commit converts the plugin to default to use sshfs in slave mode [1]
by default. Now the plugin will launch sftp-server on the host machine
and sshfs inside of the guest and wire them together over a host->guest
ssh connection. This means that we no longer have to figure out how to
connect back to the host, authenticate with the host, or even have an ssh
daemon running on the host at all. This greatly simplifies the users life.

For cases where the user was actually connecting to a third party
machine (i.e. not connecting back to the vagrant host) a user can
still do this by specifying :ssh_host in the synced folder options.

Closes #11

[1] https://github.com/libfuse/sshfs/commit/b3af91b123af607611fed00a3cc801ebc639d9ff