sshfs_mount: remove Process.kill() on failed mount
With [1] we needed to figure out a solution for Windows that was
different than for Linux. After doing some testing I have found
that we don't really need the Process.kill() calls; even without
it, the processes on Windows/Mac/Linux get cleaned up properly if
the mount fails.

Fixes #21

[1] https://github.com/dustymabe/vagrant-sshfs/issues/21