tests: don't test mounting user's home dir in tests
Testing the user's home dir is problemativ when testing vagrant in
vagrant using nested virtualization. This is because we essentially
end up trying to mount /home/vagrant from the first host over top of
/home/vagrant in the second host. We'll just stil with a simple mount
of './' for now and then check to make sure the Vagrantfile that was
used to bring up the machine is present.