sshfs_forward_mount: Don't inherit all file handles by default on Windows
We are passing in filehandles to the call to createprocess on windows, if
no filehandle is passed then no files get inherited by default, but if any
filehandle is passed to a spawned process (as we are doing) then all files
that are set as inheritable will get inherited. It just so happens that all
handles are inheritable by default.

See bInheritHandles from https://msdn.microsoft.com/en-us/library/windows/desktop/ms682425(v=vs.85).aspx

The java process is getting hung up because stdout and stderr are shared
with the spawned processes. If we set those two procs as not inheritable
by default then we don't have this problem. However, as a good practice
we'll not only set stdout and stderr as not inheritable, but all of the
other open file handles as well.

Fixes #41