sshfs_forward_mount: best effort for uninheriting
In 6f285cd we made it so that we would uninherit all filehandles by
default on windows. Some users have reported that this operation
is erroring because `The parameter is incorrect.`. See #52

In this commit I am making the uninheriting operation best effort.
The rationale is that if a file handle was not able to be set to
uninheritable then it probably wasn't one that would get inherited
in the first place. The code will now print out a warning, but
continue on.