add sudo to readlink and test -e commands
In b3eae35 and a4abb8a we added some attempt to resolve the
path of symbolic links and also a check to make sure a file path
exists before checking it. The commands we used for that don't run
privileged by default, which causes failures if you try to mount
under a directory owned by root.

Fixes #65