ssh-config: handle cases where there is no IdentityFile
In some cases there may be no IdentityFile you are using to ssh into
the guest (one example is using AWS provider and the remote AMI already
has the user's SSH key embedded in it by the cloud provider). Handle
this case.

Thanks to Michael P Reilly (https://github.com/arcege) for providing
this patch in the issue (#78).

Fixes #78.