Fixes for host path checking
For some time there have been some issues related to differing
behavior for "path validation" whether a user was executing
`vagrant sshfs` directly vs `vagrant up`. If calling `vagrant sshfs`
the host mount in the Vagrantfile wasn't checked by the code within [1]
that requires the host folder to exist on the Vagrant host machine,
but if calling `vagrant up` it was checked. We don't necessarily want
this checked because we could be using an arbitrary mount where
the actual sshfs host is not the Vagrant host. In this commit
we override the config during the validation [1] to get it to pass
validation and then we fix it back after.

Fixes #59
Fixes #72
Fixes #82

[1] https://github.com/hashicorp/vagrant/blob/030a08f869be0571e649a5372ce8b14ba752ba6c/plugins/kernel_v2/config/vm.rb#L658-L660