guest: redhat: use yum to install epel-release
Originally the epel-release rpm was not in the main repos for CentOS
and thus we needed to grab them directly from a known url. In el7 they
started putting them in the main repos and I just confirmed that
epel-release is in the main repos for el6 as well so there is no reason
not to use yum now, which will use the mirrors to pull in epel-release.

Fixes #101
Fixes #108