guest: redhat: reset back to pre-EL8 modifications
This is a revert of cd31a47 ac732ac and de0dfd7. In those commits we
were trying to support CentOS and RHEL 8 better, but the changes ended
up being mostly useful only for CentOS. Now that CentOS has it's own
flavor detection in Vagrant and centos_8 and rhel_8 are properly
supported [1] we definitely don't need de0dfd7 and we can also move
the redhat module back to just supporting RHEL since dd02e09 handles
CentOS for us.

[1] https://github.com/hashicorp/vagrant/commit/5104d075bdeb8670db7aff6c8247506e6ca28a63