add support for owner, group, and mount_options
There are some toplevel options that can be provided
to synced_folders plugins that include `owner`, `group`
and `mount_options` [1] . Let's try to add support for
them here.

The owner/group settings haven't really been an issue
for us I think because we pass in `-o allow_other` to
sshfs, but we had a request for support so let's try
to comply.

Closes https://github.com/dustymabe/vagrant-sshfs/issues/127

[1] https://github.com/hashicorp/vagrant/blob/2c3397c46851ef29a3589bf3214a3eee12da8484/website/content/docs/synced-folders/basic_usage.mdx#options